message update, i18n, core/modules/CooldownDowntime.js modification. Changelog update.
