+++ locale/de/messages.json
@@ -57,7 +57,8 @@
   "ast.lightspeed.title": "Lichtgeschwindigkeit",
   "ast.lucid-dreaming.suggestion.content": "Behalte <0/> für besseres MP-Management auf Cooldown.",
   "ast.lucid-dreaming.suggestion.why": "Ungefähr {_usesMissed, plural, one {# Cast} other {# Casts}} von Klartraum {_usesMissed, plural, one {wurde} other {wurden}} verpasst, da es für insgesamt {0} auf Bereitschaft lag.",
-  "ast.ogcd-downtime.divination": "",
+  "ast.ogcd-downtime.divination.description": "",
+  "ast.ogcd-downtime.divination.name": "",
   "ast.overheal.requirement.all": "Overheals (Insgesamt)",
   "ast.overheal.requirement.earthly-star": "Overheals (Irdischer Stern):",
   "ast.overheal.requirement.hot": "Overheals (HoTs):",
