+++ locale/en/messages.json
@@ -19,7 +19,7 @@
   "ast.celestial-intersection.suggestion.content": "Use <0/> more frequently. Frequent uses can heal or mitigate a large amount of damage over the course of a fight, potentially resulting in fewer required healing GCDs.",
   "ast.celestial-intersection.suggestion.why": "About {usesMissed} uses of <0/> were missed by holding it for at least a total of {0}.",
   "ast.celestial-opposition.suggestion.content.diurnal": "Use <0/> more frequently. In <1/>, the heal and regen combined add up to the same potency of a <2/> on each player it reaches. Trusting the regens to top off the party HP will save MP and GCDs on healing.",
-  "ast.celestial-opposition.suggestion.content.noct": "Use <0/> more frequently. In <1/>, the shield is the same potency as those from <2/>, so it can save MP and GCDs on casting it. Additionally, the shields last 30 seconds so it can be casted much earlier than the damage to be shielded and the cooldown can be ready again earlier.",
+  "ast.celestial-opposition.suggestion.content.noct": "Use <0/> more frequently. In <1/>, the shield is the same potency as from <2/>, so it can save MP and GCDs casting it. Since shields last 30 seconds it can be cast much earlier than incoming damage and allow the cooldown to refresh sooner.",
   "ast.celestial-opposition.suggestion.why": "About {usesMissed} uses of <0/> were missed by holding it for at least a total of {0}.",
   "ast.dots.requirement.uptime.name": "<0/> uptime",
   "ast.dots.rule.description": "Combust III makes up a good portion of your damage. Aim to keep this DoT up at all times. It also can be used to weave (<0/>) and manage cards, or maneuver around without dropping GCD uptime.",
@@ -57,7 +57,8 @@
   "ast.lightspeed.title": "Lightspeed",
   "ast.lucid-dreaming.suggestion.content": "Keep <0/> on cooldown for better MP management.",
   "ast.lucid-dreaming.suggestion.why": "{_usesMissed, plural, one {# use} other {# uses}} of Lucid Dreaming were missed by holding it for at least a total of {0}.",
-  "ast.ogcd-downtime.divination": "<0/> provides Astrologian with a strong amount of raid DPS when used together with arcanum. Damage percentage bonuses stack multiplicatively, so it's most optimal to stack it with cards from <1/>. Try to time the usage to match high output phases of other party members.",
+  "ast.ogcd-downtime.divination.description": "<0/> provides Astrologian with a strong amount of raid DPS when stacked together with arcanum. Damage percentage bonuses stack multiplicatively, so it's most optimal to stack it with cards from <1/>. Try to time the usage to match raid buffs and high output phases of other party members - it's more important to use it on time rather than hold it in an attempt to have 3 different seals.",
+  "ast.ogcd-downtime.divination.name": "Use Divination",
   "ast.overheal.requirement.all": "Overheal (all sources)",
   "ast.overheal.requirement.earthly-star": "Overheal (Earthly Star):",
   "ast.overheal.requirement.hot": "Overheal (HoT):",
