+++ locale/ja/messages.json
@@ -57,7 +57,8 @@
   "ast.lightspeed.title": "ライトスピード",
   "ast.lucid-dreaming.suggestion.content": "",
   "ast.lucid-dreaming.suggestion.why": "",
-  "ast.ogcd-downtime.divination": "",
+  "ast.ogcd-downtime.divination.description": "",
+  "ast.ogcd-downtime.divination.name": "",
   "ast.overheal.requirement.all": "",
   "ast.overheal.requirement.earthly-star": "",
   "ast.overheal.requirement.hot": "",
