+++ locale/ko/messages.json
@@ -57,7 +57,8 @@
   "ast.lightspeed.title": "광속 시전",
   "ast.lucid-dreaming.suggestion.content": "효율적인 마나 관리를 위해 <0/>를 상시 사용하십시오.",
   "ast.lucid-dreaming.suggestion.why": "자각몽을 최소 {0} 이상 보류 하여 {_usesMissed, plural, other {# 회}}의 사용기회를 잃어버렸습니다.",
-  "ast.ogcd-downtime.divination": "",
+  "ast.ogcd-downtime.divination.description": "",
+  "ast.ogcd-downtime.divination.name": "",
   "ast.overheal.requirement.all": "오버 힐 (전체)",
   "ast.overheal.requirement.earthly-star": "오버힐 (지상의 별):",
   "ast.overheal.requirement.hot": "오버 힐 (지속 힐):",
