+++ locale/zh/messages.json
@@ -57,7 +57,8 @@
   "ast.lightspeed.title": "光速",
   "ast.lucid-dreaming.suggestion.content": "为了你的魔力，请不要让<0/>空转。",
   "ast.lucid-dreaming.suggestion.why": "由于 CD 空转了 {0}，你大约损失了 {_usesMissed} 次醒梦。",
-  "ast.ogcd-downtime.divination": "",
+  "ast.ogcd-downtime.divination.description": "",
+  "ast.ogcd-downtime.divination.name": "",
   "ast.overheal.requirement.all": "过量治疗（所有来源）",
   "ast.overheal.requirement.earthly-star": "过量治疗（地星）：",
   "ast.overheal.requirement.hot": "过量治疗（HoT）：",
